moved entity description classes to proper file
